move commands calling code
